cron_t4/requirements.txt
streamlit
altair
pandas
requests
mysql-connector-python
